Added permit command, bugfixes, changed help
I also changed the version in the pom.xml. Here is how versioning is gonna work from now on. NO MORE PUSHING TO MASTER. Only push to bleeding and when you do add 1 to the last portion of the version, for example the next push will be beta-v0.01.02.  When we merge with master the middle number will increment. the next merges version will be beta-v0.02.01. Let's try to stick to this as much as possible, thanks!

+++ src/me/jewsofhazard/pcmrbot/database/Database.java
@@ -25,6 +25,7 @@ import java.sql.Statement;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import me.jewsofhazard.pcmrbot.MyBotMain;
 import me.jewsofhazard.pcmrbot.util.Options;
 
 public class Database {
@@ -240,7 +241,10 @@ public class Database {
 	}
 
 	public static boolean delModerator(String moderator, String channelNoHash) {
-		return executeUpdate(String.format("DELETE FROM %s.%sMods WHERE userID=\'%s\'", DATABASE, channelNoHash, moderator));
+		if(MyBotMain.isDefaultMod(moderator, channelNoHash)) {
+			return executeUpdate(String.format("DELETE FROM %s.%sMods WHERE userID=\'%s\'", DATABASE, channelNoHash, moderator));
+		}
+		return false;
 	}
 
 	public static boolean delAutoReply(String channelNoHash, String keywords) {
